+++ app/resources/views/partials/scribes/nav.blade.php
@@ -4,6 +4,7 @@
             <div class="col-md-12 col-md-12">
                 <div class="navbar-collapse">
                     <ul class="nav navbar-nav scribes-menu">
+                        <li class="{{ Route::is('scribes.overview') ? 'active' : null }}"><a href="{{ route('scribes.overview') }}">Overview</a></li>
                         <li class="{{ Route::is('scribes.races') ? 'active' : null }}"><a href="{{ route('scribes.races') }}">Races</a></li>
                         <li class="{{ Route::is('scribes.construction') ? 'active' : null }}"><a href="{{ route('scribes.construction') }}">Construction</a></li>
                         <li class="{{ Route::is('scribes.espionage') ? 'active' : null }}"><a href="{{ route('scribes.espionage') }}">Espionage</a></li>