+++ src/Http/Controllers/ScribesController.php
@@ -14,9 +14,13 @@ use OpenDominion\Models\Race;
 
 class ScribesController extends AbstractController
 {
+    public function getOverview()
+    {
+        return view('pages.scribes.overview');
+    }
+
     public function getRaces()
     {
-
         $races = collect(Race::orderBy('name')->get())->groupBy('alignment')->toArray();
         return view('pages.scribes.races', [
             'goodRaces' => $races['good'],